+++ src/Polarion/Client/PolarionClient_GetSpacesAsync.cs
@@ -12,7 +12,7 @@ namespace Polarion;
 public partial class PolarionClient : IPolarionClient
 {    
     [RequiresUnreferencedCode("Uses WCF services which require reflection")]
-    public async Task<Result<List<string>>> GetDocumentSpacesAsync()
+    public async Task<Result<List<string>>> GetSpacesAsync()
     {
         try
         {